+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggplot2 (development version)
 
+* Fixed regression where `geom_area()` didn't draw panels with single groups 
+  when `stat = "align"` (@teunbrand, #6680)
 * Fixed regression where `position_stack(vjust)` was ignored when there are
   only single groups (#6692)
 * Fixed bug where `NA` handling in `geom_path()` was ignoring panels (@teunbrand, #6533)
@@ -26,6 +28,10 @@
   insistently. Now they contribute only as fallback labels (@teunbrand, #6616)
 * Fixed regression where empty arguments to colour/fill scale caused errors
   (@jmbarbone, #6710)
+* The `theme(panel.widths, panel.heights)` setting attempts to preserve the
+  plot's aspect ratio when only one of the two settings is given, and the plot 
+  has a single panel (@teunbrand, #6701).
+* Fixed axis misplacement in `coor_radial()` when labels are blank (@teunbrand, #6574)
 
 # ggplot2 4.0.0